requirements.txt
faker==25.9.1
pandas==2.2.2
scikit-learn==1.5.2
statsmodels==0.14.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
